% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstEr.R
\name{CalcLL}
\alias{CalcLL}
\title{wrapper for Fortran function to calculate total likelihood}
\usage{
CalcLL(Er_hat, GenoM, Parents, DupsV)
}
\arguments{
\item{Er_hat}{length 3 vector with genotyping error rates}

\item{GenoM}{Genotype matrix}

\item{Parents}{Pedigree, already converted to rownumbers in GenoM}

\item{DupsV}{vector with duplicate samples, already converted to rownumbers}
}
\value{
the negative of the total log10-likelihood
}
\description{
function to be optimised
}
\keyword{internal}
