% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcRped.R
\name{CalcRped}
\alias{CalcRped}
\title{Calculate Pedigree Relatedness}
\usage{
CalcRped(Pedigree, OUT = "DF")
}
\arguments{
\item{Pedigree}{dataframe with columns id-dam-sire.}

\item{OUT}{desired output format, 'M' for matrix or 'DF' for dataframe with
columns IID1 - IID2 - R.ped.}
}
\value{
A matrix or dataframe.
}
\description{
Morph pedigree into a \pkg{kinship2} compatible format and use
  \code{\link[kinship2]{kinship}} to calculate kinship coefficients;
  relatedness = 2*kinship.
}
