% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{CheckParams}
\alias{CheckParams}
\title{Check if input parameters are valid}
\usage{
CheckParams(PARAM)
}
\arguments{
\item{PARAM}{list with input parameters}
}
\value{
Nothing except errors, warnings, and messages
}
\description{
Check if input parameter value is of the proper kind and value.
}
\keyword{internal}
