% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Conf_griffin}
\alias{Conf_griffin}
\title{Example output from estimating confidence probabilities: griffins}
\format{
a list, see \code{\link{sequoia}}
}
\usage{
data(Conf_griffin)
}
\description{
Example output of \code{\link{EstConf}}, with the inferred
  pedigree in \code{\link{SeqOUT_griffin}} used as reference pedigree.
}
\examples{
\dontrun{
Conf_griffin <- EstConf(Pedigree = SeqOUT_griffin$Pedigree,
                        LifeHistData = LH_griffin,
                        args.sim = list(nSnp = 400, SnpError = 0.001,
                                        ParMis=0.4),
                        args.seq = list(Module = 'ped', Err=0.001),
                        nSim = 20,
                        nCores = 5,
                        quiet = TRUE)
}
}
\seealso{
\code{\link{Ped_griffin}}, \code{\link{Geno_griffin}},
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
