% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstEr.R
\name{ErrPerSNP}
\alias{ErrPerSNP}
\title{wrapper for Fortran function to estimate error rate for each SNP}
\usage{
ErrPerSNP(Er_hat, GenoM, Parents, DupsV)
}
\arguments{
\item{Er_hat}{length 3 vector with genotyping error rates}

\item{GenoM}{Genotype matrix}

\item{Parents}{Pedigree, already converted to rownumbers in GenoM}

\item{DupsV}{vector with duplicate samples, already converted to rownumbers}
}
\value{
a matrix with 3 columns: for each SNP the probabilities (observed given
  actual) hom|other hom, het|hom, and hom|het.
}
\description{
WARNING: not very precise, especially not for low error rates.
}
\keyword{internal}
