% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Geno_griffin}
\alias{Geno_griffin}
\title{Example genotype file: Griffins}
\format{
A genotype matrix with 142 rows (individuals) and 200 columns (SNPs).
  Each SNP is coded as 0/1/2 copies of the reference allele, with -9 for
  missing values. Ids are stored as rownames.
}
\usage{
data(Geno_griffin)
}
\description{
Simulated genotype data from Pedigree \code{\link{Ped_griffin}}
}
\seealso{
\code{\link{SimGeno}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
