% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Geno_HSg5}
\alias{Geno_HSg5}
\title{Example genotype file: 'HSg5'}
\format{
A genotype matrix with 920 rows (ids) and 200 columns (SNPs). Each
  SNP is coded as 0/1/2 copies of the reference allele, with -9 for missing
  values. Ids are stored as rownames.
}
\usage{
data(Geno_HSg5)
}
\description{
Simulated genotype data for all* individuals in Pedigree
  \code{\link{Ped_HSg5}} (*: with 40% of parents are non-genotyped).
}
\examples{
\dontrun{
# this output was created as follows:
Geno_HSg5 <- SimGeno(Ped = Ped_HSg5, nSnp = 200, ParMis=0.4,
                     CallRate = 0.9, SnpError = 0.005)
}
}
\seealso{
\code{\link{LH_HSg5}, \link{SimGeno}, \link{SeqOUT_HSg5}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
