% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindFamilies.R
\name{GetDescendants}
\alias{GetDescendants}
\title{Get descendants}
\usage{
GetDescendants(id, Pedigree)
}
\arguments{
\item{id}{id of the individual}

\item{Pedigree}{dataframe with columns id - parent1 - parent2; only the
first 3 columns will be used.}
}
\value{
a list with as first element \code{id}, second offspring, third
 grand-offspring, etc.. Each element is a vector with ids, the first three
 elements are named, the rest numbered.
}
\description{
get all descendants of an individual
}
