% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Inherit}
\alias{Inherit}
\title{Inheritance patterns}
\format{
An array with the following dimensions:
\describe{
  \item{d1}{type: autosomal, x-chromosome, y-chromosome, or mtDNA}
  \item{d2}{offspring sex: female, male, or unknown}
  \item{d3}{offspring genotype: aa (0), aA (1), Aa (1), or AA (2)}
  \item{d4}{mother genotype}
  \item{d5}{father genotype}
}
}
\usage{
data(Inherit)
}
\description{
Inheritance patterns used by SimGeno for non-autosomal SNPs,
  identical to those in Inherit.xlsx
}
\seealso{
\code{\link{SimGeno}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{inherit}
\keyword{sequoia}
