% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{LH_HSg5}
\alias{LH_HSg5}
\title{Example life history file: 'HSg5'}
\format{
A data frame with 1000 rows and 3 variables:
\describe{
\item{ID}{Female IDs start with 'a', males with 'b'; the next 2 numbers give
  the generation number (00 -- 05), the last 3 numbers the individual ID
  number (runs continuously across all generations)}
\item{Sex}{1 = female, 2 = male}
\item{BirthYear}{from 2000 (generation 0, founders) to 2005} }
}
\usage{
data(LH_HSg5)
}
\description{
This is the life history file associated with
  \code{\link{Ped_HSg5}}, which is \strong{Pedigree II} in the paper.
}
\references{
Huisman, J. (2017) Pedigree reconstruction from SNP data:
  Parentage assignment, sibship clustering, and beyond. Molecular Ecology
  Resources 17:1009--1024.
}
\seealso{
\code{\link{Ped_HSg5} \link{sequoia}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
