% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{MaybeRel_griffin}
\alias{MaybeRel_griffin}
\title{Example output from check for relatives: griffins}
\format{
a list with 2 dataframes, 'MaybePar' and 'MaybeTrio'. See
  \code{\link{GetMaybeRel}} for further details.
}
\usage{
data(MaybeRel_griffin)
}
\description{
Example output of a check for parent-offspring pairs and
  parent-parent-offspring trios with \code{\link{GetMaybeRel}}, with
  \code{\link{Geno_griffin}} as input (simulated from
  \code{\link{Ped_griffin}}).
}
\examples{
\dontrun{
MaybeRel_griffin <- GetMaybeRel(GenoM = Geno_griffin, Err=0.001,
                                Module = 'par')
}
}
\seealso{
\code{\link{SeqOUT_griffin}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
