% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{MergeFill}
\alias{MergeFill}
\title{Special Merge}
\usage{
MergeFill(df1, df2, by, overwrite = FALSE, ...)
}
\arguments{
\item{df1}{first dataframe (lowest priority if \code{overwrite=TRUE}).}

\item{df2}{second dataframe (highest priority if \code{overwrite=TRUE}).}

\item{by}{columns used for merging, required.}

\item{overwrite}{If FALSE (the default), NA's in df1 are replaced by values
from df2. If TRUE, all values in df1 are overwritten by values from df2,
except where df2 has NA.}

\item{...}{additional arguments to merge, such as \code{all}.}
}
\description{
As regular merge, but combine data from columns with the same
 name.
}
\keyword{internal}
