% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MkAgePrior.R
\name{MkAPdefault}
\alias{MkAPdefault}
\title{Make default 1/0 ageprior}
\usage{
MkAPdefault(MinP, MaxP, Disc)
}
\arguments{
\item{MinP}{Minimum age of dams, sires}

\item{MaxP}{Maximum age of dams, sires}

\item{Disc}{Discrete generations? TRUE/FALSE/NULL}
}
\description{
Create ageprior matrix based on min and max age of parents.
}
\keyword{internal}
