% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{MkFortParams}
\alias{MkFortParams}
\title{PARAM to FortPARAM}
\usage{
MkFortParams(PARAM, fun = "main")
}
\arguments{
\item{PARAM}{list with input parameters.}

\item{fun}{function from which \code{\link{MkFortParams}} is called,
determines which elements are included in the output list.}
}
\value{
A list with elements
  \item{Ng}{Integer, number of individuals}
  \item{SpecsInt}{8 integers:
    \itemize{
      \item nSnp
      \item MaxMismatchV; DUP - OH - ME
      \item MaxSibshipSize
      \item Complx, 0=mono, 1=simp, 2=full
      \item quiet, -1=verbose, 0=FALSE, 1=TRUE
      \item nAgeCl, nrow(AgePriors)
    }
  }
  \item{SpecsDbl}{2 double precision numbers:
    \itemize{
      \item Tfilter  (< 0)
      \item Tassign  (> 0)
    }
  }
  \item{ErrM}{double, 3x3 matrix passed as length-9 vector}
  \item{SpecsIntMkPed}{\code{fun='main'} only
    \itemize{
      \item MaxSibIter
      \item AgeEffect, 0=no, 1=yes, 2=extra
      \item CalcLLR, 0=FALSE, 1=TRUE
      \item Herm, 0=no, 1= dam/sire distinction, 2=no dam/sire distinction
    }
  }
  \item{SpecsIntAmb}{\code{fun='mayberel'} only
    \itemize{
      \item ParSib  1=par, 2=ped
      \item nAmbMax
    }
  }
}
\description{
Convert list \code{PARAM} into a list with integer-only and
  double-only vectors, to be passed to Fortran.
}
\keyword{internal}
