% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{mtSame2Dif}
\alias{mtSame2Dif}
\title{Check and recode mtSame matrix}
\usage{
mtSame2Dif(mtSame = NULL, gID = NULL)
}
\arguments{
\item{mtSame}{matrix indicating whether individuals (might) have the same
mitochondrial haplotype (1),  or definitely not (0). Not all individuals
need to be included and order is not important, may not be square.}

\item{gID}{rownames of `GenoM`}
}
\value{
square gID x gID matrix indicating whether individuals have
  definitely a different mitochondrial haplotype (1), or (possibly) the same
  (0).
}
\description{
Recode to 1=different, 0=same
}
\keyword{internal}
