% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{ParamToSpecs}
\alias{ParamToSpecs}
\title{PARAM to Specs}
\usage{
ParamToSpecs(PARAM, TimeStart, ErrFlavour)
}
\arguments{
\item{PARAM}{list with input parameters.}

\item{TimeStart}{time at which \code{\link{sequoia}} run was started.}

\item{ErrFlavour}{character name or function.}
}
\value{
The 1-row \code{Specs} dataframe.
}
\description{
Convert list \code{PARAM} into 1-row dataframe \code{Specs}.
  Only to be called by \code{\link{sequoia}}.
}
\keyword{internal}
