% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqListSummary.R
\name{plot.seq}
\alias{plot.seq}
\title{Plot Summary Overview of sequoia Output}
\usage{
\method{plot}{seq}(SeqSum, Pedigree, Panels = "all")
}
\arguments{
\item{SeqSum}{list output from \code{\link{SummarySeq}}.}

\item{Pedigree}{dataframe with at least id, dam and sire in columns 1-3,
respectively. If columns with parental LLRs and/or Mendelian errors are
present, these will be plotted as well.}

\item{Panels}{character vector with panel(s) to plot. Choose from 'all',
'G.parents', 'D.parents', 'sibships', 'LLR', 'OH'.}
}
\description{
visualise the numbers of assigned parents, sibship sizes, and
  parental LLRs
}
\keyword{internal}
