% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPstats.R
\name{PlotSnpStats}
\alias{PlotSnpStats}
\title{plot SnpStats results}
\usage{
PlotSnpStats(OUT)
}
\arguments{
\item{OUT}{output from \code{SnpStats}}
}
\value{
plots
}
\description{
scatter plots and histograms of allele frequency, missingness,
  and estimated genotyping error, across SNPs
}
\keyword{internal}
