% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimGeno.R
\name{SelectNotSampled}
\alias{SelectNotSampled}
\title{select non-genotyped parents}
\usage{
SelectNotSampled(Ped, ParMis)
}
\arguments{
\item{Ped}{pedigree, after PedPolish()}

\item{ParMis}{single number or vector length two with proportion of parents
with fully missing genotype}
}
\value{
vector with genotype matrix row numbers of non-sampled individuals
}
\description{
select non-genotyped parents
}
\keyword{internal}
