% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{SeqOUT_griffin}
\alias{SeqOUT_griffin}
\title{Example output from pedigree inference: griffins}
\format{
a list, see \code{\link{sequoia}}
}
\usage{
data(SeqOUT_griffin)
}
\description{
Example output of a sequoia run including sibship clustering,
  with \code{\link{Geno_griffin}} as input (simulated from
  \code{\link{Ped_griffin}}).
}
\examples{
\dontrun{
SeqOUT_griffin <- sequoia(GenoM = Geno_griffin,
                          LifeHistData = LH_griffin,
                          Module = 'ped')
}
}
\seealso{
\code{\link{sequoia}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
