% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{SpecsToParam}
\alias{SpecsToParam}
\title{Specs to PARAM}
\usage{
SpecsToParam(Specs, ErrM = NULL, ErrFlavour = NULL, dimGeno = NULL, ...)
}
\arguments{
\item{Specs}{1-row dataframe, element of \code{\link{sequoia}} output list.}

\item{...}{other objects to append to the list, such as \code{ErrM} and
\code{quiet}.}
}
\value{
A named list.
}
\description{
Convert 1-row dataframe \code{Specs} into list \code{PARAM},
  optionally including various other objects in the list.
}
\keyword{internal}
