% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparePeds.R
\name{Vcomp}
\alias{Vcomp}
\title{Compare two vectors}
\usage{
Vcomp(Infrd, Simld, SNPd)
}
\arguments{
\item{Infrd}{vector of inferred sibs}

\item{Simld}{vector of true sibs}

\item{SNPd}{character vector with IDs of genotyped individuals}
}
\value{
a named numeric vector of length 4, with the total length of Simld,
  the length of the intersect of the two vectors, the number occurring in
  Infrd but not Simld ('err'), and the number occuring in Simld but not
  Infrd ('missed').
}
\description{
Compare a vector with inferred sibs to a vector of `true' sibs
}
\keyword{internal}
