% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPedComp.R
\name{VennSquares}
\alias{VennSquares}
\title{Square Venn diagram}
\usage{
VennSquares(count, BL = c(0, 0), COL, withText = TRUE, withLegend = FALSE)
}
\arguments{
\item{count}{a length 5 named vector: 'Total', 'Match', 'Mismatch',
'P1only', and 'P2only'.}

\item{BL}{a length 2 vector with coordinates of bottom-mid of Ped1 square.}

\item{COL}{a length 4 character vector with colours, named 'Match',
'Mismatch', 'Ped1', 'Ped2'.}

\item{withText}{logical, add count to each rectangle.}

\item{withLegend}{logical, add legend at the bottom of the plot.}
}
\description{
Draw Venn diagram with squares, with match/mismatch in
  overlapping area.
}
\seealso{
\code{\link{PlotPedComp}}
}
\keyword{internal}
