% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{LH_griffin}
\alias{LH_griffin}
\title{Example lifehistory data: griffins}
\format{
A data frame with 200 rows and 3 variables (ID, Sex, BirthYear)
}
\usage{
data(LH_griffin)
}
\description{
Example lifehistory data for griffin pedigree
}
\seealso{
\code{\link{Ped_griffin}}, \code{\link{SeqOUT_griffin}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
